AWS constants: move to cmk.utils
