+++ cmk/utils/aws_constants.py
