+++ tests/unit/cmk/special_agents/agent_aws/agent_aws_fake_clients.py
@@ -4,7 +4,7 @@ import abc
 import random
 import six
 
-from cmk.special_agents.agent_aws_types import AWSEC2InstTypes
+from cmk.utils.aws_constants import AWSEC2InstTypes
 
 #   .--entities------------------------------------------------------------.
 #   |                             _   _ _   _                              |
